Rename the main directory and fix imports
