--- sudoku/api.py
@@ -1,1 +1,0 @@
-from sudoku.hb6d_board import HB6DBoard as Board  # noqa: F401
